refactor: view 클래스의 경우 QueryBuilder로 변경
