+++ src/main/java/persistence/sql/ddl/DropQueryBuilder.java
@@ -1,12 +1,11 @@
 package persistence.sql.ddl;
 
-import persistence.common.EntityClazz;
 import persistence.sql.QueryBuilder;
 
 public class DropQueryBuilder implements QueryBuilder {
     @Override
     public String build(Object obj) {
-        EntityClazz entityClazz = new EntityClazz(obj.getClass());
-        return entityClazz.getDropQuery() + ";";
+        EntityQueryBuilder entityQueryBuilder = new EntityQueryBuilder(obj.getClass());
+        return entityQueryBuilder.getDropQuery() + ";";
     }
 }
